requirements.txt
Django==5.2.6
psycopg2-binary==2.9.10
celery==5.5.3
django-celery-results==2.6.0
djangorestframework==3.16.1
django-redis==5.4.0
redis==5.0.1
